[STF] Fix error in link_check and a bunch of other warnings (#4024)
* Fix error in linker_test and a bunch of other warnings

* Fix formatting

---------

+++ cudax/examples/stf/linear_algebra/07-cholesky.cu
@@ -174,7 +174,7 @@ public:
     // Fill blocks by blocks
     for (size_t colb = 0; colb < nt; colb++)
     {
-      int low_rowb = sym_matrix ? colb : 0;
+      size_t low_rowb = sym_matrix ? colb : 0;
       for (size_t rowb = low_rowb; rowb < mt; rowb++)
       {
         // Each task fills a block
